Move new_resume over to being resume and change CI test suite target to resume
